// src/global/dto/detalle-venta-dto.ts
export class DetalleVentaDTO {
    id: number | null = null;
    idventa: number | null = null;
    monto: number | null = null;
    cantidad: number | null = null;
    subtotal: number | null = null;
    idservicio: number | null = null;
    servicio: string | null = null;
    idgrupo: number | null = null;
    grupo: string | null = null;
    porcentajeiva: number | null = null;
    montoiva: number | null = null;
    idcuota: number | null = null;
    descripcion: string | null = null;
    idsuscripcion: number | null = null;
    eliminado: boolean | null = null;
}